requirements.txt
requests==2.32.3
mcp==1.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
